add convenience accessor for revocation keys
It may not be obvious for users of the API that we can get the list of
revoker keys for a given key from the (private) _signature list, so
add a convenient accessor.

This is not really useful right now because it will raise a
NotImplementedError if any such signature is found, but will become
actually quite useful once #198 lands.

This is part of the process to make revocation checks easier in #225.