keyserver search and get support
Basic primitives to search keyservers and fetch associated keys.

There are a lot of issues with this code, but it is a first proof of
concept. The first issue is we do not return a meaningful
datastructure in `search`: we should really return a NamedTuple or,
better, existing sturctures like PGPKey and PGPUID, but I couldn't
figure out how to attach all data fields. Also, key fields like
primary key expiration date are missing.

More importantly, we do not check at all what we get from the
keyserver. We blindly load this in a PGP key and hope for the
best. Maybe that's alright: we do not know what the caller wants to do
and maybe that should be left to the caller to do proper
verification. But we may want to check if the key we get matches the
fingerprint we load and raise otherwise.

There's also the question of how to do HTTP requests. Here we
introduce a dependency on the `requests` module, which works well in
my tests, but maybe a little advanced. It *will* allow us to easily do
HTTPS checks eventually, however. We also don't support hkp://
protocols - maybe that's just a requests `adapter` we should be
writing, actually.

There's also the question of how to run HTTP requests in the test
suite. In another project, I have used the `betamax` module to record
HTTP queries and replay them so that we could have a stable testbed
while offline, but also keep the network inputs controlled and
constant so that even if they change (e.g. a key is changed, the test
suite is unaffected.

Finally, I have no idea where to put this code. I put it in `types.py`
because that is where existing `from_blob` and `from_file` functions
were, so I figured it would make sense to put them there. But I really
don't know.