test: stop not being expired, on static date
tests/test_03_armor.py::block_attrs was setting:
"block.is_expired = False"
Which was correct at the time the test was created, but it's no
longer true based on the the expiration date in revochiio.asc
("datetime(2018, 9, 12, 1, 0, 59))")
While there's not mocking on times, tests should not relay on
dates.
If there are more tests depending on dates, they will fail in the
future too.