fixed failing tests
Commit e163a0cf22735d639ecaaeab4cd6c903acb9a5c5 made creation and
expiration date relative to current time which does not match content in
revochiio.asc. Reverted commit and asserted the key as expired.

test_10_exceptions.py: test_encrypt_sessionkey_wrongtype did not have a
wrong sessionkey type and therefore was not raising TypeError.

pytest.raises() message parameter is deprecated
https://github.com/pytest-dev/pytest/pull/4539

EllipticCurveOID does not have a name attribute, the name of the curve
is stored in EllipticCurveOID.curve.name