User IDs: simplify, treat as UTF-8 strings
Over in #271, we discussed simplifying the lower-level UserID packet
object, and moving the <name,comment,email> splitting logic up into
the higher-layer PGPUID API.

This patch does that work, and also exposes an additional
PGPUID.userid property, which is the full UTF-8 string.

If this gets merged, it will obsolete #271.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>