PGPKey._sign: include hashed Issuer Fingerprint subpacket by default.
GnuPG has been including hashed Issuer Fingerprint subpackets by
default in signatures since 2016-10-08.  This subpacket makes it
possible to distinguish between a bad signature and a signature that
we happen to not have the issuer's key for.

We add it here for normal signature, certifications, and revocations,
unless the signer explicitly requests that it not be included.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>