Added PGPMessage.message to docs
Added the undocumented .message API call to the documentation, as it's essential for reading compressed and encrypted messages (since encrypted messages are currently required to be compressed), as well as automatically removing PGP headers. 

Also added it to the examples, as you can't read any of them if they're compressed without reimplementing the compression scheme used in PGPMessage.message

This api call is used in tests/test_05_actions.py, though it isn't documented anywhere else.