PGPObject: add type signatures
Note that bytes_to_text and text_to_bytes no longer accept None as input.

When they would accept None as input, that meant that their output could also be None
which in turn made it harder to vet the type of the output when the input was
clearly not None.