PGPUID: add type annotations
Note that you used to be able to use the or operator to append UserID and
UserAttribute packets indefinitely, but only the first packet would take effect.
all the other ones would be silently ignored.

the current code is simpler, but instead ignores all the previous packets,
and uses the last one.