explicitly document the ability to sign encrypted messages
I'm deeply skeptical about the sense or utility of this, but
it is at least documenting the current behavior more adequately.

The resulting messages (when signed) form this packet sequence:
SIG PKESK SEIPD

This is technically valid in the formal OpenPGP message grammar
but it's unclear what it is supposed to mean.