Parse embedded signature without assuming version
This will be important if a non-v4 signature gets embedded.

It might end up being an opaque signature, if PGPy can't read the
version, but it shouldn't try to force a SignatureV4 if it is not
actually version 4.

This change also isolates the parsing of the embedded signature so that
it can only consume bytes within the enclosing subpacket.